Pushing changes to enable Rails 3 / Rspec 2 compatibility
