+++ Rakefile
@@ -1,5 +1,6 @@
 require 'rake'
 require 'rake/testtask'
+require 'yaml'
 begin
   require 'jeweler'
   Jeweler::Tasks.new do |s|
@@ -15,15 +16,10 @@ rescue LoadError
   puts "Jeweler not available. Install it with: sudo gem install technicalpickels-jeweler -s http://gems.github.com"
 end
 
-require 'spec/rake/spectask'
-Spec::Rake::SpecTask.new(:spec) do |spec|
-  spec.libs << 'lib' << 'spec'
-  spec.spec_files = FileList['spec/**/*_spec.rb']
-end
+require 'rspec/core/rake_task'
 
-Spec::Rake::SpecTask.new(:rcov) do |spec|
-  spec.libs << 'lib' << 'spec'
-  spec.pattern = 'spec/**/*_spec.rb'
+RSpec::Core::RakeTask.new(:spec)
+RSpec::Core::RakeTask.new(:rcov) do |spec|
   spec.rcov = true
 end
 task :default => :spec
