+++ spec/resources/config/database.yml
@@ -1,3 +1,3 @@
 sqlite3:
   :adapter: sqlite3
-  :dbfile: vendor/plugins/object_daddy/spec/tmp/object_daddy.db
+  :database: vendor/plugins/object_daddy/spec/tmp/object_daddy.db
