--- spec/spec.opts
@@ -1,5 +1,0 @@
---colour
---format
-s
---loadby
-mtime
